Modularize app by splitting recipe generation, PDF export, translation, nutrition, and history logic into separate scripts
