+++ history_utils.py
@@ -1,0 +1,18 @@
+import os
+import json
+
+def load_recipe_history(history_file="recipe_history.json"):
+    if os.path.exists(history_file):
+        try:
+            with open(history_file, "r", encoding="utf-8") as f:
+                return json.load(f)
+        except Exception:
+            return []
+    return []
+
+def save_recipe_history(history, history_file="recipe_history.json"):
+    try:
+        with open(history_file, "w", encoding="utf-8") as f:
+            json.dump(history, f, ensure_ascii=False, indent=2)
+    except Exception:
+        pass
