+++ recipe_utils.py
@@ -1,0 +1,33 @@
+def extract_recipe_name(recipe_text):
+    """
+    Extracts the recipe name from the generated text.
+    Prioritizes lines starting with "1. CREATIVE RECIPE NAME:" or "RECIPE NAME:".
+    """
+    lines = recipe_text.strip().split('\n')
+    for line in lines:
+        cleaned_line = line.strip()
+        if not cleaned_line:
+            continue
+        name_part = ""
+        if cleaned_line.lower().startswith("1. creative recipe name:"):
+            name_part = cleaned_line.split(":", 1)[1].strip()
+        elif cleaned_line.lower().startswith("creative recipe name:"):
+            name_part = cleaned_line.split(":", 1)[1].strip()
+        elif cleaned_line.lower().startswith("1. recipe name:"):
+            name_part = cleaned_line.split(":", 1)[1].strip()
+        elif cleaned_line.lower().startswith("recipe name:"):
+            name_part = cleaned_line.split(":", 1)[1].strip()
+        elif cleaned_line.lower().startswith("1."):
+            potential_name = cleaned_line[len("1."):].strip()
+            if len(potential_name.split()) < 10 and not any(kw in potential_name.lower() for kw in ["ingredient", "step", "prep", "cook", "serving", "total time", "description"]):
+                name_part = potential_name
+            else:
+                continue
+        elif not name_part and len(lines) > 0 and lines[0].strip() == cleaned_line:
+            if len(cleaned_line.split()) < 10 and not any(kw in cleaned_line.lower() for kw in ["ingredient", "step", "prep", "cook", "serving", "total time", "description"]):
+                name_part = cleaned_line
+        if name_part:
+            name = name_part.replace("*", "").replace("#", "").strip(":- ")
+            if name:
+                return name[:150]
+    return "Untitled Recipe"
